+++ HISTORY.rst
@@ -3,6 +3,12 @@
 Release History
 ---------------
 
+0.3.0 (2014-xx-xx)
+++++++++++++++++++
+- Allow keyword arguments in functions (e.g. `plot([1,2,3], linewidth=2))`).
+  These are translated to ("prop", value) arguments to the function.
+
+
 0.2.0 (2014-08-14)
 ++++++++++++++++++
 - Streamline API to mirror Oct2Py 2.0.0